Fix another bug (when provide no extra argument for keyword, crash)
