+++ src/main/java/Duke.java
@@ -26,9 +26,6 @@ public class Duke {
         }
     }
 
-    public static void convertList(ArrayList<Duke.Task> tasks) {
-
-    }
     public static String formatStr(String str) {
         String returnstr =  ">>>>>>>>>>>>>>>>>>>>>>>>>>>>>\n"
                             + str + "\n"
@@ -36,50 +33,6 @@ public class Duke {
         return returnstr;
     }
 
-    public static class Task {
-        protected boolean mark;
-        protected String content;
-
-        public Task(String content) {
-            this.content = content;
-            this.mark = false;
-        }
-
-        public Task(String content, boolean alternative) {
-            this.content = content;
-            this.mark = alternative;
-        }
-
-        public boolean getMark() {
-            return this.mark;
-        }
-
-        public void setMark() {
-            this.mark = !this.mark;
-            String outputStr;
-            if (mark == true) {
-                outputStr = "NICE! You finished this: \n"
-                        + "[" + markSign(this.mark) + "] " + this.content;
-            } else {
-                outputStr = "Ok, you have undone this: \n"
-                        + "[" + markSign(this.mark) + "] " + this.content;
-            }
-            System.out.println(formatStr(outputStr));
-        }
-
-        public String markSign(boolean markBool) {
-            if(markBool == true) return "X";
-            else return " ";
-        }
-
-        public String toString() {
-            return ". [" + markSign(this.mark) + "] " + this.content;
-        }
-
-        public String printRecord() {
-            return this.toString();
-        }
-    }
 
 }
