+++ src/main/java/Todo.java
@@ -1,4 +1,4 @@
-public class Todo extends Duke.Task {
+public class Todo extends Task {
     public Todo(String content) {
         super(content.substring(5));
     }
