+++ src/test/java/TodoTest.java
@@ -1,0 +1,8 @@
+package task.Todo;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class TodoTest {
+}
